backend/requirements.txt
fastapi
uvicorn
dashscope
python-dotenv
pydantic
neo4j
